refactor: apply additional changes requested in pull request review
