+++ src/pages/registration-page/ui/RegistrationPage.module.scss
@@ -1,29 +1,33 @@
 .page-wrapper {
   @apply flex justify-center bg-zinc-200 min-h-screen;
 
-  @media screen and (min-width: 672px) {
+  @media screen and (min-width: 712px) {
     @apply w-screen;
   }
 }
 
 .registration-page {
-  @apply min-w-[380px] px-5 flex justify-center self-center py-8 rounded-lg bg-white border-b-[10px] border-[#46a358];
+  @apply min-w-[380px] px-5 flex justify-center self-center py-6 rounded-lg bg-white border-b-[10px] border-[#46a358];
 
-  @media screen and (min-width: 792px) {
+  @media screen and (min-width: 832px) {
     @apply w-[792px];
+  }
+
+  @media screen and (max-width: 712px) {
+    @apply my-5;
   }
 }
 
 .form-wrapper {
   @apply max-w-[632px];
 
-  @media screen and (max-width: 671px) {
+  @media screen and (max-width: 712px) {
     @apply max-w-[352px];
   }
 }
 
 .form-header {
-  @apply mb-8 flex justify-center gap-3;
+  @apply mb-4 flex justify-center gap-3;
 }
 
 .form-header__link {
@@ -41,7 +45,7 @@
 .form-text {
   @apply text-sm font-["Cera_Pro_Regular"];
 
-  @media screen and (max-width: 671px) {
+  @media screen and (max-width: 712px) {
     @apply ml-[26px];
   }
 }
