+++ src/shared/BlockDivider/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './ui/BlockDivider';
