+++ src/shared/BlockDivider/ui/BlockDivider.module.scss
@@ -1,4 +1,4 @@
-.form-divider {
+.block-divider {
   @apply mt-8 flex align-middle;
 }
