+++ src/widgets/BillingAddress/ui/BillingAddress.module.scss
@@ -1,7 +1,7 @@
 .from-text__right {
   @apply text-sm font-["Cera_Pro_Regular"] flex justify-end;
 
-  @media screen and (max-width: 671px) {
+  @media screen and (max-width: 712px) {
     @apply mr-[26px];
   }
 }
