+++ src/widgets/BillingAddress/ui/BillingAddress.tsx
@@ -1,18 +1,17 @@
-import Address from '../Address';
-import FormDivider from '../FormDivider';
+import Address from '@/features/Address';
+import BlockDivider from '@/shared/BlockDivider';
 import styles from './BillingAddress.module.scss';
 
 const BillingAddress = () => {
   return (
     <>
-      <FormDivider text={'Billing address'} />
+      <BlockDivider text={'Billing address'} />
       <Address />
       <label className={styles['from-text__right']} htmlFor='bill-default'>
         Set as default billing address
         <input
           className={styles['form-checkbox-right']}
           type='checkbox'
-          name=''
           id='bill-default'
         />
       </label>
