--- src/widgets/FormDivider/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './FormDivider';
