+++ src/widgets/UserDetails/ui/UserDetails.module.scss
