+++ src/widgets/UserDetails/ui/UserDetails.tsx
